requirements.txt
aenum==3.1.15
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
dataclasses-json==0.6.7
deepgram-sdk==3.8.0
deprecation==2.1.0
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.12.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
Jinja2==3.1.5
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.29
langchain-ollama==0.2.2
langsmith==0.2.10
MarkupSafe==3.0.2
marshmallow==3.24.1
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
numpy==2.2.1
ollama==0.4.5
opencv-python==4.10.0.84
orjson==3.10.13
outcome==1.3.0.post0
packaging==24.2
pip==24.0
propcache==0.2.1
psutil==6.1.1
PyAudio==0.2.14
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
PySocks==1.7.1
python-dotenv==1.0.1
pyvts==0.3.3
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
selenium==4.27.1
setuptools==75.7.0
sniffio==1.3.1
sortedcontainers==2.4.0
sympy==1.13.1
tenacity==9.0.0
torch==2.5.1
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
typing-inspect==0.9.0
urllib3==2.3.0
websocket-client==1.8.0
websockets==13.1
wsproto==1.2.0